LoggingService: Move "download as file" logic, into separate module (#745)
The extractLogFile() method of LoggingService contains code that
handles creating and downloading the log file. This code is not
strongly related to logging.

Let's move this file-downloading logic into a separate
module to simplify and increase the modularity of LoggingService.
